00-bootcamp-project/pyproject.toml
[tool.poetry]
name = "00-bootcamp-project"
version = "0.1.0"
description = ""
authors = ["Kiattiphum <83391695+KiattiphumSuw@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
requests = "2.31.0"
pysftp = "0.2.9"
psycopg2 = "2.9.9"
google-cloud-bigquery = "3.17.2"
python-dotenv = "^1.0.1"
google-cloud-storage = "^2.14.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
